More improvements to SensorState. This is all over the place now.Added more documentation and the option to get single data points immediately. Also made SensorContainer private due to potential sync issues.

+++ FtcRobotController/src/main/java/lib/TwoWheelDrive.java
@@ -59,8 +59,9 @@ public class TwoWheelDrive implements DriveTrain {
         rightMotor.setPower(0);
     }
 
-    public void turnWithGyro(double power, int degrees) {
-
+    public void turnWithGyro(double power, double degrees) {
+//        double goal = gyro.getRotation + degrees;
+//        while (gyro.getRotation() - initial < )
     }
 
 //    public void turnWithGyro(float power, double heading)
